fix(tests): harden tests against hardcoded paths, state leaks and weak assertions
- Replace /tmp and relative tmp/ directories with tempfile helpers.
- Extract repeated HTTP server setup into a context manager.
- Restore mocked vllm module and environment after test.
- Strengthen prepare_model/get_model/free_models assertions.
- Use real absolute paths instead of POSIX-only literals.
